migrate image uploads to Supabase
