+++ package.json
@@ -12,7 +12,7 @@
   "author": "Bernoulli Mukuna",
   "license": "ISC",
   "dependencies": {
-    "aws-sdk": "^2.1692.0",
+    "@supabase/supabase-js": "^2.45.4",
     "axios": "^1.7.7",
     "bcrypt": "^5.1.1",
     "cookie-parser": "^1.4.6",
@@ -25,7 +25,6 @@
     "moment": "^2.30.1",
     "mongoose": "^8.5.1",
     "multer": "^1.4.5-lts.2",
-    "multer-s3": "^3.0.1",
     "nodemailer": "^6.9.14",
     "nodemon": "^3.1.4",
     "stripe": "^16.6.0",
